requirements.txt
fastapi~=0.85.0
uvicorn
jinja2
python-multipart
git+https://github.com/openai/whisper.git
opencv-contrib-python
pandas
python-docx

numpy
torch
tqdm
more-itertools
transformers>=4.19.0
ffmpeg-python==0.2.0
pydantic
